+++ src/account_readers/dvl_readable.rs
@@ -1,4 +1,5 @@
 use std::error::Error;
+use async_trait::async_trait;
 use solana_program::pubkey::Pubkey;
 use crate::dvl_client::dvl_client::DvlClient;
 use crate::accounts::client::client_account::client_account::ClientAccount;
@@ -18,25 +19,26 @@ pub struct DvlClientParams<'a> {
     pub signer_account_params: Option<&'a SignerAccountParams<'a>>,
 }
 
+#[async_trait]
 pub trait DvlReadable {
     type DvlReadParams<'a>;
 
-    fn get_public_key<'a>(client: &DvlClient, params: &Self::DvlReadParams<'a>) -> Result<Box<Pubkey>, Box<dyn Error>>
+    async fn get_public_key<'a>(client: &DvlClient, params: &Self::DvlReadParams<'a>) -> Result<Box<Pubkey>, Box<dyn Error>>
         where
             Self: Sized;
 
-    fn read<'a>(client: &DvlClient, params: &Self::DvlReadParams<'a>) -> Result<Box<Self>, Box<dyn Error>>
+    async fn read<'a>(client: &DvlClient, params: &Self::DvlReadParams<'a>) -> Result<Box<Self>, Box<dyn Error>>
         where
             Self: Sized;
 
-    fn read_by_public_key(
+    async fn read_by_public_key(
         client: &DvlClient,
         public_key: &Pubkey,
     ) -> Result<Box<Self>, Box<dyn Error>>
         where
             Self: Sized + DevolAccount + Copy
     {
-        let mut rpc_data = client.rpc_client.get_account(public_key)?;
+        let mut rpc_data = client.rpc_client.get_account(public_key).await?;
         let account = Self::from_account_basic(
             public_key,
             &mut rpc_data,