+++ src/account_readers/read_worker_account.rs
@@ -1,4 +1,5 @@
 use std::error::Error;
+use async_trait::async_trait;
 use solana_program::pubkey::Pubkey;
 use crate::dvl_client::dvl_client::DvlClient;
 use crate::account_readers::dvl_readable::{DvlReadable, DvlIndexParam};
@@ -6,17 +7,19 @@ use crate::accounts::all_workers::all_workers_account::AllWorkersAccount;
 use crate::accounts::devol_indexed_account::DevolIndexedAccount;
 use crate::accounts::worker::worker_account::WorkerAccount;
 
+#[async_trait]
 impl DvlReadable for WorkerAccount {
     type DvlReadParams<'a> = DvlIndexParam;
-    fn get_public_key<'a>(client: &DvlClient, params: &Self::DvlReadParams<'a>) -> Result<Box<Pubkey>, Box<dyn Error>> where Self: Sized {
-        let workers_account = client.get_account::<AllWorkersAccount>(()).unwrap();
+
+    async fn get_public_key<'a>(client: &DvlClient, params: &Self::DvlReadParams<'a>) -> Result<Box<Pubkey>, Box<dyn Error>> where Self: Sized {
+        let workers_account = client.get_account::<AllWorkersAccount>(()).await?;
         let worker = workers_account.workers[params.id as usize];
         Ok(Box::from(worker.address))
     }
 
-    fn read<'a>(client: &DvlClient, params: &Self::DvlReadParams<'a>) -> Result<Box<Self>, Box<dyn Error>> where Self: Sized {
-        let public_key = &*Self::get_public_key(client, params)?;
-        let mut rpc_data = client.rpc_client.get_account(public_key)?;
+    async fn read<'a>(client: &DvlClient, params: &Self::DvlReadParams<'a>) -> Result<Box<Self>, Box<dyn Error>> where Self: Sized {
+        let public_key = &*Self::get_public_key(client, params).await?;
+        let mut rpc_data = client.rpc_client.get_account(public_key).await?;
         let account =  Self::from_account(
             public_key,
             &mut rpc_data,
@@ -34,19 +37,19 @@ mod tests {
     use crate::tests::tests::setup_devol_client;
     use std::error::Error;
 
-    #[test]
-    fn test_read_worker_account_by_index() -> Result<(), Box<dyn Error>> {
+    #[tokio::test]
+    async fn test_read_worker_account_by_index() -> Result<(), Box<dyn Error>> {
         let client = setup_devol_client();
-        let _worker_0 = client.get_account::<WorkerAccount>(DvlIndexParam { id: 0 })?;
+        let _worker_0 = client.get_account::<WorkerAccount>(DvlIndexParam { id: 0 }).await?;
         Ok(())
     }
 
-    #[test]
-    fn test_read_worker_account_by_public_key() -> Result<(), Box<dyn Error>> {
+    #[tokio::test]
+    async fn test_read_worker_account_by_public_key() -> Result<(), Box<dyn Error>> {
         let client = setup_devol_client();
-        let mints_account = client.get_account::<AllWorkersAccount>(())?;
+        let mints_account = client.get_account::<AllWorkersAccount>(()).await?;
         let pubkey = &mints_account.workers[0].address;
-        let _worker_0 = client.get_account_by_public_key::<WorkerAccount>(pubkey)?;
+        let _worker_0 = client.get_account_by_public_key::<WorkerAccount>(pubkey).await?;
         Ok(())
     }
 }