+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-05-06
+
+### Changed
+- Switched over to async approach
+- Start next pool params constructor now receive f32 array as pool bounds params instead of i32
+- DvlClient receive Pubkey instead of &str
+
 ## [0.1.3] - 2024-05-03
 
 ### Fixed